Add iOS boilerplate to keep pub publish happy
Despite being an Android-only plugin, pub publish has taken it upon
itself to reject any apps that do not include boilerplate iOS
scaffolding, regardless of whether iOS is supported or not.

This is supposed to be 'fixed' in the upcoming 1.20.0 release, but it
makes little sense to constrain the plugin to a release that hasn't even
happened yet.

Rip this back out once Flutter behaves the way it should have from the
beginning.